requirements.txt
flask
flask-restful
flask-sqlalchemy
flask-marshmallow
marshmallow-sqlalchemy
PyMySQL
flask-jwt-extended
flask-cors
werkzeug
bcrypt
cryptography
